Add 4 sysctls: vol_offset_line, vol_scale_line, vol_offset_hp, vol_scale_hp.
These sysctls make it possible to efficently switch output from Line-Out to (Rear)Headphones and
vice versa. For headphones with their resistance close enough to 600Ω default settings must be
OK. For more common headphones with resistance < 50-60Ω lesser values of vol_scale_hp may be
required. Sum of vol_offset_* and vol_scale_* must not exceed 255.

The reason for this:

When you switch output to different destination, the actual volume of the sound may be changed,
in other words "mixer vol 60" on headphones may be loud enough, and "mixer vol 60" on line-out
may be complete silence. User may play with newly added sysctls to achieve the moment when the
same settings of mixer sound the same on different outputs.

The values of these settings are preserved between boots with xonarstate startup script.